Moved to_csvy function to writers submodule
