+++ tardis/visualization/widgets/custom_abundance.py
@@ -24,6 +24,7 @@ from tardis.io.model_reader import (
     read_uniform_abundances,
     parse_csv_abundances,
 )
+from tardis.io.writers.custom_abundance import CustomAbundanceCSVYWriter
 from tardis.util.base import atomic_number2element_symbol, quantity_linspace
 from tardis.visualization.tools.convergence_plot import transition_colors
 from tardis.visualization.widgets.util import debounce
@@ -288,60 +289,6 @@ class CustomAbundanceWidgetData:
             abundance=abundance,
             velocity=velocity,
         )
-
-
-class CustomYAML(yaml.YAMLObject):
-    """A custom YAML object generated by required properties."""
-
-    def __init__(
-        self, name, d_time_0, i_time_0, v_inner_boundary, v_outer_boundary
-    ):
-        """Initialize CustomYAML object with given properties.
-
-        Parameters
-        ----------
-        name : str
-            Name of the YAML file.
-        d_time_0 : astropy.units.quantity.Quantity
-            Initial time for the density in the model.
-        i_time_0 : astropy.units.quantity.Quantity
-            Initial time for isotope decay. Set to 0 for no isotopes.
-        v_inner_boundary : astropy.units.quantity.Quantity
-            Velocity of the inner boundary.
-        v_outer_boundary : astropy.units.quantity.Quantity
-            Velocity of the outer boundary.
-        """
-        self.name = name
-        self.model_density_time_0 = d_time_0
-        self.model_isotope_time_0 = i_time_0
-        self.tardis_model_config_version = "v1.0"
-        self.datatype = {}
-        self.datatype["fields"] = []
-        self.v_inner_boundary = v_inner_boundary
-        self.v_outer_boundary = v_outer_boundary
-
-    def create_fields_dict(self, elements):
-        """Create a dictionary to store the items in 'fields' part.
-
-        Parameters
-        ----------
-        elements : list of str
-            A list of elements or isotopes' symbols.
-        """
-        for i in range(len(elements) + 2):
-            field = {}
-
-            if i == 0:
-                field["name"] = "velocity"
-                field["unit"] = "km/s"
-            elif i == 1:
-                field["name"] = "density"
-                field["unit"] = "g/cm^3"
-            else:
-                field["name"] = elements[i - 2]
-                field["desc"] = f"fractional {elements[i-2]} abundance"
-
-            self.datatype["fields"].append(field)
 
 
 class CustomAbundanceWidget:
@@ -1132,7 +1079,7 @@ class CustomAbundanceWidget:
         path = self.input_path.value
         overwrite = self.ckb_overwrite.value
 
-        self.to_csvy(path, overwrite)
+        CustomAbundanceCSVYWriter.to_csvy(self, path, overwrite)
 
     def rbs_single_apply_eventhandler(self, obj):
         """Switch to single shell editing mode. Triggered if the
@@ -1379,90 +1326,6 @@ class CustomAbundanceWidget:
                 self.error_view,
             ]
         )
-
-    @error_view.capture(clear_output=True)
-    def to_csvy(self, path, overwrite):
-        """Output CSVY file on the specified path.
-
-        Parameters
-        ----------
-        path : str
-            Output path.
-        overwrite : bool
-            True if overwriting, False otherwise.
-        """
-        posix_path = Path(path)
-        posix_path = posix_path.with_suffix(".csvy")
-
-        if posix_path.exists() and not overwrite:
-            raise FileExistsError(
-                "The file already exists. Click the 'overwrite' checkbox to overwrite it."
-            )
-        else:
-            self.write_yaml_portion(posix_path)
-            self.write_csv_portion(posix_path)
-
-    @error_view.capture(clear_output=True)
-    def write_yaml_portion(self, path):
-        """Write the YAML portion of the output file.
-
-        Parameters
-        ----------
-        path : pathlib.PosixPath
-        """
-        name = path.name
-        d_time_0 = self.data.density_t_0
-        i_time_0 = self.input_i_time_0.value * u.day
-        custom_yaml = CustomYAML(
-            name,
-            d_time_0,
-            i_time_0,
-            self.data.velocity[0],
-            self.data.velocity[-1],
-        )
-        custom_yaml.create_fields_dict(self.data.elements)
-
-        with path.open("w") as f:
-            yaml_output = yaml.dump(custom_yaml, sort_keys=False)
-
-            # Add YAML delimiter
-            yaml_output_snippets = yaml_output.split("\n")
-            yaml_output_snippets[0] = yaml_output_snippets[-1] = YAML_DELIMITER
-            yaml_output = "\n".join(yaml_output_snippets) + "\n"
-
-            f.write(yaml_output)
-
-        print("Saved Successfully!")
-
-    def write_csv_portion(self, path):
-        """Write the CSV portion of the output file.
-
-        Parameters
-        ----------
-        path : pathlib.PosixPath
-        """
-        try:
-            data = self.data.abundance.T
-            data.columns = self.data.elements
-            first_row = [0] * self.no_of_elements
-            data.loc[-1] = first_row
-            data.index += 1  # shifting index
-            data.sort_index(inplace=True)
-
-            formatted_v = pd.Series(self.data.velocity.value).apply(
-                lambda x: "%.3e" % x
-            )
-            # Make sure velocity is within the boundary.
-            formatted_v[0] = self.data.velocity.value[0]
-            formatted_v[-1] = self.data.velocity.value[-1]
-
-            density = self.data.density
-            data.insert(0, "velocity", formatted_v)
-            data.insert(1, "density", density)
-
-            data.to_csv(path, mode="a", index=False)
-        except pd.errors.EmptyDataError:
-            data = None
 
     @classmethod
     def from_csvy(cls, fpath):
